+++ src/dom/route.test.ts
@@ -26,9 +26,9 @@ it("provides absolute path", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.path, "/test/child">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.path, "/test/child/grand">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.path).toEqual("/test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toEqual("/test/child/grand");
+    expect(TEST_ROUTE.path).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.path).toStrictEqual("/test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toStrictEqual("/test/child/grand");
 });
 
 it("provides absolute path with optional segments", () => {
@@ -40,9 +40,9 @@ it("provides absolute path with optional segments", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.path, "/test/optional?/child/optional-child?">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.path, "/test/optional?/child/optional-child?/grand">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/test/optional?");
-    expect(TEST_ROUTE.CHILD.path).toEqual("/test/optional?/child/optional-child?");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toEqual("/test/optional?/child/optional-child?/grand");
+    expect(TEST_ROUTE.path).toStrictEqual("/test/optional?");
+    expect(TEST_ROUTE.CHILD.path).toStrictEqual("/test/optional?/child/optional-child?");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toStrictEqual("/test/optional?/child/optional-child?/grand");
 });
 
 it("provides absolute path with optional dynamic segments", () => {
@@ -54,9 +54,9 @@ it("provides absolute path with optional dynamic segments", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.path, "/test/:optional?/child/:optional-child?">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.path, "/test/:optional?/child/:optional-child?/grand">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/test/:optional?");
-    expect(TEST_ROUTE.CHILD.path).toEqual("/test/:optional?/child/:optional-child?");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toEqual("/test/:optional?/child/:optional-child?/grand");
+    expect(TEST_ROUTE.path).toStrictEqual("/test/:optional?");
+    expect(TEST_ROUTE.CHILD.path).toStrictEqual("/test/:optional?/child/:optional-child?");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toStrictEqual("/test/:optional?/child/:optional-child?/grand");
 });
 
 it("preserves intermediate stars in absolute path", () => {
@@ -68,9 +68,9 @@ it("preserves intermediate stars in absolute path", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.path, "/test/child/*">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.path, "/test/child/*/grand">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.path).toEqual("/test/child/*");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toEqual("/test/child/*/grand");
+    expect(TEST_ROUTE.path).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.path).toStrictEqual("/test/child/*");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toStrictEqual("/test/child/*/grand");
 });
 
 it("preserves optional intermediate stars in absolute path", () => {
@@ -82,9 +82,9 @@ it("preserves optional intermediate stars in absolute path", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.path, "/test/*?/child/*?">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.path, "/test/*?/child/*?/grand">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/test/*?");
-    expect(TEST_ROUTE.CHILD.path).toEqual("/test/*?/child/*?");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toEqual("/test/*?/child/*?/grand");
+    expect(TEST_ROUTE.path).toStrictEqual("/test/*?");
+    expect(TEST_ROUTE.CHILD.path).toStrictEqual("/test/*?/child/*?");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toStrictEqual("/test/*?/child/*?/grand");
 });
 
 it("provides relative path", () => {
@@ -96,9 +96,9 @@ it("provides relative path", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.relativePath, "test/child">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.relativePath, "test/child/grand">>(true);
 
-    expect(TEST_ROUTE.relativePath).toEqual("test");
-    expect(TEST_ROUTE.CHILD.relativePath).toEqual("test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toEqual("test/child/grand");
+    expect(TEST_ROUTE.relativePath).toStrictEqual("test");
+    expect(TEST_ROUTE.CHILD.relativePath).toStrictEqual("test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toStrictEqual("test/child/grand");
 });
 
 it("provides relative path with optional segments", () => {
@@ -110,9 +110,9 @@ it("provides relative path with optional segments", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.relativePath, "test/dynamic?/child/:dynamic-param?">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.relativePath, "test/dynamic?/child/:dynamic-param?/grand">>(true);
 
-    expect(TEST_ROUTE.relativePath).toEqual("test/dynamic?");
-    expect(TEST_ROUTE.CHILD.relativePath).toEqual("test/dynamic?/child/:dynamic-param?");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toEqual("test/dynamic?/child/:dynamic-param?/grand");
+    expect(TEST_ROUTE.relativePath).toStrictEqual("test/dynamic?");
+    expect(TEST_ROUTE.CHILD.relativePath).toStrictEqual("test/dynamic?/child/:dynamic-param?");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toStrictEqual("test/dynamic?/child/:dynamic-param?/grand");
 });
 
 it("removes intermediate stars from relative path", () => {
@@ -124,9 +124,9 @@ it("removes intermediate stars from relative path", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.relativePath, "test/child/*">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.relativePath, "test/child/grand">>(true);
 
-    expect(TEST_ROUTE.relativePath).toEqual("test");
-    expect(TEST_ROUTE.CHILD.relativePath).toEqual("test/child/*");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toEqual("test/child/grand");
+    expect(TEST_ROUTE.relativePath).toStrictEqual("test");
+    expect(TEST_ROUTE.CHILD.relativePath).toStrictEqual("test/child/*");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toStrictEqual("test/child/grand");
 });
 
 it("removes multiple intermediate stars from relative path", () => {
@@ -138,9 +138,9 @@ it("removes multiple intermediate stars from relative path", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.relativePath, "test/child/*">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.relativePath, "test/child/grand">>(true);
 
-    expect(TEST_ROUTE.relativePath).toEqual("test/*");
-    expect(TEST_ROUTE.CHILD.relativePath).toEqual("test/child/*");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toEqual("test/child/grand");
+    expect(TEST_ROUTE.relativePath).toStrictEqual("test/*");
+    expect(TEST_ROUTE.CHILD.relativePath).toStrictEqual("test/child/*");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toStrictEqual("test/child/grand");
 });
 
 it("removes multiple optional intermediate stars from relative path", () => {
@@ -152,9 +152,9 @@ it("removes multiple optional intermediate stars from relative path", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.relativePath, "test/child/*?">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.relativePath, "test/child/grand">>(true);
 
-    expect(TEST_ROUTE.relativePath).toEqual("test/*?");
-    expect(TEST_ROUTE.CHILD.relativePath).toEqual("test/child/*?");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toEqual("test/child/grand");
+    expect(TEST_ROUTE.relativePath).toStrictEqual("test/*?");
+    expect(TEST_ROUTE.CHILD.relativePath).toStrictEqual("test/child/*?");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toStrictEqual("test/child/grand");
 });
 
 it("allows empty segment at the beginning of the route", () => {
@@ -170,13 +170,13 @@ it("allows empty segment at the beginning of the route", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.relativePath, "child">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.relativePath, "child/grand">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/");
-    expect(TEST_ROUTE.CHILD.path).toEqual("/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toEqual("/child/grand");
+    expect(TEST_ROUTE.path).toStrictEqual("/");
+    expect(TEST_ROUTE.CHILD.path).toStrictEqual("/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toStrictEqual("/child/grand");
 
-    expect(TEST_ROUTE.relativePath).toEqual("");
-    expect(TEST_ROUTE.CHILD.relativePath).toEqual("child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toEqual("child/grand");
+    expect(TEST_ROUTE.relativePath).toStrictEqual("");
+    expect(TEST_ROUTE.CHILD.relativePath).toStrictEqual("child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toStrictEqual("child/grand");
 });
 
 it("allows empty segment in the middle of the route", () => {
@@ -192,13 +192,13 @@ it("allows empty segment in the middle of the route", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.relativePath, "test">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.relativePath, "test/grand">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.path).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toEqual("/test/grand");
+    expect(TEST_ROUTE.path).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.path).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toStrictEqual("/test/grand");
 
-    expect(TEST_ROUTE.relativePath).toEqual("test");
-    expect(TEST_ROUTE.CHILD.relativePath).toEqual("test");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toEqual("test/grand");
+    expect(TEST_ROUTE.relativePath).toStrictEqual("test");
+    expect(TEST_ROUTE.CHILD.relativePath).toStrictEqual("test");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toStrictEqual("test/grand");
 });
 
 it("allows empty segment at the end of the route", () => {
@@ -214,13 +214,13 @@ it("allows empty segment at the end of the route", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.relativePath, "test/child">>(true);
     assert<IsExact<typeof TEST_ROUTE.CHILD.GRANDCHILD.relativePath, "test/child">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.path).toEqual("/test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toEqual("/test/child");
+    expect(TEST_ROUTE.path).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.path).toStrictEqual("/test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.path).toStrictEqual("/test/child");
 
-    expect(TEST_ROUTE.relativePath).toEqual("test");
-    expect(TEST_ROUTE.CHILD.relativePath).toEqual("test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toEqual("test/child");
+    expect(TEST_ROUTE.relativePath).toStrictEqual("test");
+    expect(TEST_ROUTE.CHILD.relativePath).toStrictEqual("test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.relativePath).toStrictEqual("test/child");
 });
 
 it("allows implicit path params", () => {
@@ -232,9 +232,9 @@ it("allows implicit path params", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], { id: string }>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], { id: string }>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ id: "42" })).toEqual("/test/child/42");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: "24" })).toEqual("/test/child/24/grand");
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ id: "42" })).toStrictEqual("/test/child/42");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: "24" })).toStrictEqual("/test/child/24/grand");
 });
 
 it("allows explicit path params", () => {
@@ -246,9 +246,9 @@ it("allows explicit path params", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], { id: number }>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], { id: number }>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ id: 42 })).toEqual("/test/child/42");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: 24 })).toEqual("/test/child/24/grand");
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ id: 42 })).toStrictEqual("/test/child/42");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: 24 })).toStrictEqual("/test/child/24/grand");
 });
 
 it("allows to mix explicit and implicit path params", () => {
@@ -260,9 +260,9 @@ it("allows to mix explicit and implicit path params", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], { id: number; value: string }>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], { id: number; value: string }>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ id: 42, value: "foo" })).toEqual("/test/child/42/foo");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: 24, value: "bar" })).toEqual("/test/child/24/bar/grand");
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ id: 42, value: "foo" })).toStrictEqual("/test/child/42/foo");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: 24, value: "bar" })).toStrictEqual("/test/child/24/bar/grand");
 });
 
 it("allows to mix explicit and implicit path params in case of optional params", () => {
@@ -279,9 +279,9 @@ it("allows to mix explicit and implicit path params in case of optional params",
         >
     >(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ id: 42 })).toEqual("/test/child/42");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: 24, value: "bar", name: "baz" })).toEqual(
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ id: 42 })).toStrictEqual("/test/child/42");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: 24, value: "bar", name: "baz" })).toStrictEqual(
         "/test/child/24/bar/grand/baz"
     );
 });
@@ -300,9 +300,9 @@ it("allows to mix explicit and implicit path params across multiple routes", () 
         >
     >(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ id: 42, value: "foo" })).toEqual("/test/child/42/foo");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: 24, value: "bar", name: "baz" })).toEqual(
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ id: 42, value: "foo" })).toStrictEqual("/test/child/42/foo");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: 24, value: "bar", name: "baz" })).toStrictEqual(
         "/test/child/24/bar/grand/baz"
     );
 });
@@ -316,9 +316,9 @@ it("prioritizes children when mixing path params with the same name", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], { id: number; value: string }>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], { id: boolean; value: string }>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ id: 42, value: "foo" })).toEqual("/test/child/42/foo");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: false, value: "bar" })).toEqual(
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ id: 42, value: "foo" })).toStrictEqual("/test/child/42/foo");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ id: false, value: "bar" })).toStrictEqual(
         "/test/child/false/bar/grand/false"
     );
 });
@@ -332,9 +332,9 @@ it("allows implicit star path param", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], { "*"?: string }>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], Record<never, never>>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ "*": "star/param" })).toEqual("/test/child/star/param");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({})).toEqual("/test/child/grand");
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ "*": "star/param" })).toStrictEqual("/test/child/star/param");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ "*": "star/param" })).toStrictEqual("/test/child/grand");
 });
 
 it("allows implicit optional star path param", () => {
@@ -346,9 +346,9 @@ it("allows implicit optional star path param", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], { "*"?: string }>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], Record<never, never>>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ "*": "star/param" })).toEqual("/test/child/star/param");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({})).toEqual("/test/child/grand");
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ "*": "star/param" })).toStrictEqual("/test/child/star/param");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ "*": "star/param" })).toStrictEqual("/test/child/grand");
 });
 
 it("allows explicit star path param", () => {
@@ -360,9 +360,9 @@ it("allows explicit star path param", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], Record<never, never>>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], { "*"?: number }>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({})).toEqual("/test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ "*": 42 })).toEqual("/test/child/grand/42");
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({})).toStrictEqual("/test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ "*": 42 })).toStrictEqual("/test/child/grand/42");
 });
 
 it("allows explicit optional star path param", () => {
@@ -374,9 +374,9 @@ it("allows explicit optional star path param", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], Record<never, never>>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], { "*"?: number }>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({})).toEqual("/test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ "*": 42 })).toEqual("/test/child/grand/42");
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({})).toStrictEqual("/test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ "*": 42 })).toStrictEqual("/test/child/grand/42");
 });
 
 it("always treats star param as optional upon building", () => {
@@ -388,9 +388,9 @@ it("always treats star param as optional upon building", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], Record<never, never>>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], { "*"?: number }>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({})).toEqual("/test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({})).toEqual("/test/child/grand");
+    expect(TEST_ROUTE.buildPath({})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({})).toStrictEqual("/test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({})).toStrictEqual("/test/child/grand");
 });
 
 it("allows star path param in the middle of combined path", () => {
@@ -402,9 +402,9 @@ it("allows star path param in the middle of combined path", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[0], { "*"?: string }>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[0], Record<never, never>>>(true);
 
-    expect(TEST_ROUTE.buildPath({})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({ "*": "foo" })).toEqual("/test/child/foo");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({})).toEqual("/test/child/grand");
+    expect(TEST_ROUTE.buildPath({ "*": "foo" })).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({ "*": "foo" })).toStrictEqual("/test/child/foo");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({ "*": "foo" })).toStrictEqual("/test/child/grand");
 });
 
 it("allows search params", () => {
@@ -416,9 +416,9 @@ it("allows search params", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[1], Record<never, never> | undefined>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[1], { foo?: number } | undefined>>(true);
 
-    expect(TEST_ROUTE.buildPath({}, {})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({}, {})).toEqual("/test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, { foo: 1 })).toEqual("/test/child/grand?foo=1");
+    expect(TEST_ROUTE.buildPath({}, {})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({}, {})).toStrictEqual("/test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, { foo: 1 })).toStrictEqual("/test/child/grand?foo=1");
 });
 
 it("allows to mix search params across multiple routes", () => {
@@ -435,9 +435,9 @@ it("allows to mix search params across multiple routes", () => {
         >
     >(true);
 
-    expect(TEST_ROUTE.buildPath({}, {})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({}, { bar: [1, 2] })).toEqual("/test/child?bar=1&bar=2");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, { foo: 1, bar: [1, 2] })).toEqual(
+    expect(TEST_ROUTE.buildPath({}, {})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({}, { bar: [1, 2] })).toStrictEqual("/test/child?bar=1&bar=2");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, { foo: 1, bar: [1, 2] })).toStrictEqual(
         "/test/child/grand?foo=1&bar=1&bar=2"
     );
 });
@@ -451,9 +451,9 @@ it("prioritizes children when mixing search params with the same name", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[1], { foo?: number[] } | undefined>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[1], { foo?: number } | undefined>>(true);
 
-    expect(TEST_ROUTE.buildPath({}, {})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({}, { foo: [1, 2] })).toEqual("/test/child?foo=1&foo=2");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, { foo: 1 })).toEqual("/test/child/grand?foo=1");
+    expect(TEST_ROUTE.buildPath({}, {})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({}, { foo: [1, 2] })).toStrictEqual("/test/child?foo=1&foo=2");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, { foo: 1 })).toStrictEqual("/test/child/grand?foo=1");
 });
 
 it("allows implicit hash params", () => {
@@ -465,9 +465,9 @@ it("allows implicit hash params", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[2], undefined>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[2], string | undefined>>(true);
 
-    expect(TEST_ROUTE.buildPath({}, {})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({}, {})).toEqual("/test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, {}, "my-id")).toEqual("/test/child/grand#my-id");
+    expect(TEST_ROUTE.buildPath({}, {})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({}, {})).toStrictEqual("/test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, {}, "my-id")).toStrictEqual("/test/child/grand#my-id");
 });
 
 it("allows explicit hash params", () => {
@@ -479,9 +479,9 @@ it("allows explicit hash params", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[2], undefined>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[2], "foo" | "bar" | undefined>>(true);
 
-    expect(TEST_ROUTE.buildPath({}, {})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({}, {})).toEqual("/test/child");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, {}, "foo")).toEqual("/test/child/grand#foo");
+    expect(TEST_ROUTE.buildPath({}, {})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({}, {})).toStrictEqual("/test/child");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, {}, "foo")).toStrictEqual("/test/child/grand#foo");
 });
 
 it("allows mixing explicit hash params across multiple routes", () => {
@@ -495,9 +495,9 @@ it("allows mixing explicit hash params across multiple routes", () => {
         true
     );
 
-    expect(TEST_ROUTE.buildPath({}, {})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({}, {}, "baz")).toEqual("/test/child#baz");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, {}, "baz")).toEqual("/test/child/grand#baz");
+    expect(TEST_ROUTE.buildPath({}, {})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({}, {}, "baz")).toStrictEqual("/test/child#baz");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, {}, "baz")).toStrictEqual("/test/child/grand#baz");
 });
 
 it("allows mixing explicit and implicit hash params across multiple routes", () => {
@@ -509,9 +509,9 @@ it("allows mixing explicit and implicit hash params across multiple routes", () 
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildPath>[2], "baz" | undefined>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildPath>[2], string | undefined>>(true);
 
-    expect(TEST_ROUTE.buildPath({}, {})).toEqual("/test");
-    expect(TEST_ROUTE.CHILD.buildPath({}, {}, "baz")).toEqual("/test/child#baz");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, {}, "anything")).toEqual("/test/child/grand#anything");
+    expect(TEST_ROUTE.buildPath({}, {})).toStrictEqual("/test");
+    expect(TEST_ROUTE.CHILD.buildPath({}, {}, "baz")).toStrictEqual("/test/child#baz");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildPath({}, {}, "anything")).toStrictEqual("/test/child/grand#anything");
 });
 
 it("allows state params", () => {
@@ -523,9 +523,9 @@ it("allows state params", () => {
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.buildState>[0], { foo?: string }>>(true);
     assert<IsExact<Parameters<typeof TEST_ROUTE.CHILD.GRANDCHILD.buildState>[0], { foo?: string; bar?: number }>>(true);
 
-    expect(TEST_ROUTE.buildState({})).toEqual({});
-    expect(TEST_ROUTE.CHILD.buildState({ foo: "test" })).toEqual({ foo: "test" });
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildState({ foo: "test", bar: 1 })).toEqual({ foo: "test", bar: 1 });
+    expect(TEST_ROUTE.buildState({})).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.buildState({ foo: "test" })).toStrictEqual({ foo: "test" });
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.buildState({ foo: "test", bar: 1 })).toStrictEqual({ foo: "test", bar: 1 });
 });
 
 it("allows implicit path params parsing", () => {
@@ -537,9 +537,9 @@ it("allows implicit path params parsing", () => {
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.getTypedParams>, Record<never, never>>>(true);
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams>, { id: string }>>(true);
 
-    expect(TEST_ROUTE.getTypedParams({})).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedParams({})).toEqual({});
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ id: "1" })).toEqual({ id: "1" });
+    expect(TEST_ROUTE.getTypedParams({})).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedParams({})).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ id: "1" })).toStrictEqual({ id: "1" });
     expect(() => TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({})).toThrow();
 });
 
@@ -552,10 +552,10 @@ it("allows implicit optional path params parsing", () => {
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.getTypedParams>, Record<never, never>>>(true);
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams>, { id?: string }>>(true);
 
-    expect(TEST_ROUTE.getTypedParams({})).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedParams({})).toEqual({});
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ id: "1" })).toEqual({ id: "1" });
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({})).toEqual({});
+    expect(TEST_ROUTE.getTypedParams({})).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedParams({})).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ id: "1" })).toStrictEqual({ id: "1" });
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({})).toStrictEqual({});
 });
 
 it("allows explicit path params parsing", () => {
@@ -567,9 +567,9 @@ it("allows explicit path params parsing", () => {
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.getTypedParams>, Record<never, never>>>(true);
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams>, { id: number | undefined }>>(true);
 
-    expect(TEST_ROUTE.getTypedParams({ id: "1" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedParams({ id: "1" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ id: "1" })).toEqual({ id: 1 });
+    expect(TEST_ROUTE.getTypedParams({ id: "1" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedParams({ id: "1" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ id: "1" })).toStrictEqual({ id: 1 });
 });
 
 it("allows to mix path params parsing across multiple routes", () => {
@@ -586,9 +586,9 @@ it("allows to mix path params parsing across multiple routes", () => {
         >
     >(true);
 
-    expect(TEST_ROUTE.getTypedParams({ id: "1", childId: "2" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedParams({ id: "1", childId: "2" })).toEqual({ childId: 2 });
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ id: "1", childId: "2" })).toEqual({ id: 1, childId: 2 });
+    expect(TEST_ROUTE.getTypedParams({ id: "1", childId: "2" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedParams({ id: "1", childId: "2" })).toStrictEqual({ childId: 2 });
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ id: "1", childId: "2" })).toStrictEqual({ id: 1, childId: 2 });
 });
 
 it("provides untyped path params", () => {
@@ -711,9 +711,9 @@ it("allows implicit star path param parsing", () => {
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.getTypedParams>, Record<never, never>>>(true);
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams>, { "*": string }>>(true);
 
-    expect(TEST_ROUTE.getTypedParams({ "*": "foo/bar" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedParams({ "*": "foo/bar" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ "*": "foo/bar" })).toEqual({ "*": "foo/bar" });
+    expect(TEST_ROUTE.getTypedParams({ "*": "foo/bar" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedParams({ "*": "foo/bar" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ "*": "foo/bar" })).toStrictEqual({ "*": "foo/bar" });
 });
 
 it("allows implicit optional star path param parsing", () => {
@@ -739,9 +739,9 @@ it("allows explicit star path param parsing", () => {
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.getTypedParams>, Record<never, never>>>(true);
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams>, { "*": number | undefined }>>(true);
 
-    expect(TEST_ROUTE.getTypedParams({ "*": "1" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedParams({ "*": "1" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ "*": "1" })).toEqual({ "*": 1 });
+    expect(TEST_ROUTE.getTypedParams({ "*": "1" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedParams({ "*": "1" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ "*": "1" })).toStrictEqual({ "*": 1 });
 });
 
 it("allows explicit star path param parsing (with default value)", () => {
@@ -753,9 +753,9 @@ it("allows explicit star path param parsing (with default value)", () => {
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.getTypedParams>, Record<never, never>>>(true);
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams>, { "*": number }>>(true);
 
-    expect(TEST_ROUTE.getTypedParams({ "*": "" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedParams({ "*": "" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ "*": "" })).toEqual({ "*": 42 });
+    expect(TEST_ROUTE.getTypedParams({ "*": "" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedParams({ "*": "" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ "*": "" })).toStrictEqual({ "*": 42 });
 });
 
 it("doesn't throw if explicit star param is invalid", () => {
@@ -781,9 +781,9 @@ it("allows intermediate star param parsing", () => {
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.getTypedParams>, { "*": string }>>(true);
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams>, { "*": string }>>(true);
 
-    expect(TEST_ROUTE.getTypedParams({ "*": "foo/bar" })).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedParams({ "*": "foo/bar" })).toEqual({ "*": "foo/bar" });
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ "*": "foo/bar" })).toEqual({ "*": "foo/bar" });
+    expect(TEST_ROUTE.getTypedParams({ "*": "foo/bar" })).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedParams({ "*": "foo/bar" })).toStrictEqual({ "*": "foo/bar" });
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedParams({ "*": "foo/bar" })).toStrictEqual({ "*": "foo/bar" });
 });
 
 it("allows search params parsing", () => {
@@ -807,21 +807,23 @@ it("allows search params parsing", () => {
 
     const testSearchParams = createSearchParams({ arr: ["1", "2"], foo: "foo", untyped: "untyped" });
 
-    expect(TEST_ROUTE.getTypedSearchParams(testSearchParams)).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedSearchParams(testSearchParams)).toEqual({ arr: [1, 2], foo: "foo" });
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedSearchParams(testSearchParams)).toEqual({ arr: [1, 2], foo: 0 });
+    expect(TEST_ROUTE.getTypedSearchParams(testSearchParams)).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedSearchParams(testSearchParams)).toStrictEqual({ arr: [1, 2], foo: "foo" });
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedSearchParams(testSearchParams)).toStrictEqual({ arr: [1, 2], foo: 0 });
 
-    expect(TEST_ROUTE.getUntypedSearchParams(testSearchParams).getAll("arr")).toEqual(["1", "2"]);
-    expect(TEST_ROUTE.CHILD.getUntypedSearchParams(testSearchParams).getAll("arr")).toEqual([]);
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getUntypedSearchParams(testSearchParams).getAll("arr")).toEqual([]);
+    expect(TEST_ROUTE.getUntypedSearchParams(testSearchParams).getAll("arr")).toStrictEqual(["1", "2"]);
+    expect(TEST_ROUTE.CHILD.getUntypedSearchParams(testSearchParams).getAll("arr")).toStrictEqual([]);
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getUntypedSearchParams(testSearchParams).getAll("arr")).toStrictEqual([]);
 
-    expect(TEST_ROUTE.getUntypedSearchParams(testSearchParams).get("foo")).toEqual("foo");
-    expect(TEST_ROUTE.CHILD.getUntypedSearchParams(testSearchParams).get("foo")).toEqual(null);
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getUntypedSearchParams(testSearchParams).get("foo")).toEqual(null);
+    expect(TEST_ROUTE.getUntypedSearchParams(testSearchParams).get("foo")).toStrictEqual("foo");
+    expect(TEST_ROUTE.CHILD.getUntypedSearchParams(testSearchParams).get("foo")).toStrictEqual(null);
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getUntypedSearchParams(testSearchParams).get("foo")).toStrictEqual(null);
 
-    expect(TEST_ROUTE.getUntypedSearchParams(testSearchParams).get("untyped")).toEqual("untyped");
-    expect(TEST_ROUTE.CHILD.getUntypedSearchParams(testSearchParams).get("untyped")).toEqual("untyped");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getUntypedSearchParams(testSearchParams).get("untyped")).toEqual("untyped");
+    expect(TEST_ROUTE.getUntypedSearchParams(testSearchParams).get("untyped")).toStrictEqual("untyped");
+    expect(TEST_ROUTE.CHILD.getUntypedSearchParams(testSearchParams).get("untyped")).toStrictEqual("untyped");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getUntypedSearchParams(testSearchParams).get("untyped")).toStrictEqual(
+        "untyped"
+    );
 });
 
 it("throws if throwable search params are invalid", () => {
@@ -845,8 +847,8 @@ it("throws if throwable search params are invalid", () => {
 
     const testSearchParams = createSearchParams({ arr: ["1", "2"], foo: "foo", untyped: "untyped" });
 
-    expect(TEST_ROUTE.getTypedSearchParams(testSearchParams)).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedSearchParams(testSearchParams)).toEqual({ arr: [1, 2], foo: "foo" });
+    expect(TEST_ROUTE.getTypedSearchParams(testSearchParams)).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedSearchParams(testSearchParams)).toStrictEqual({ arr: [1, 2], foo: "foo" });
     expect(() => TEST_ROUTE.CHILD.GRANDCHILD.getTypedSearchParams(testSearchParams)).toThrow();
 });
 
@@ -867,9 +869,9 @@ it("allows hash parsing", () => {
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.getTypedHash>, "foo" | "bar" | undefined>>(true);
     assert<IsExact<ReturnType<typeof TEST_ROUTE.CHILD.GRANDCHILD.getTypedHash>, string | undefined>>(true);
 
-    expect(TEST_ROUTE.getTypedHash(testHash)).toEqual(undefined);
-    expect(TEST_ROUTE.CHILD.getTypedHash(testHash)).toEqual("foo");
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedHash(testHash)).toEqual("foo");
+    expect(TEST_ROUTE.getTypedHash(testHash)).toStrictEqual(undefined);
+    expect(TEST_ROUTE.CHILD.getTypedHash(testHash)).toStrictEqual("foo");
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedHash(testHash)).toStrictEqual("foo");
 });
 
 it("allows any hash parsing", () => {
@@ -885,9 +887,9 @@ it("allows any hash parsing", () => {
 
     const testHash = "#baz";
 
-    expect(TEST_ROUTE.getTypedHash(testHash)).toEqual(undefined);
-    expect(TEST_ROUTE.CHILD.getTypedHash(testHash)).toEqual(undefined);
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedHash(testHash)).toEqual("baz");
+    expect(TEST_ROUTE.getTypedHash(testHash)).toStrictEqual(undefined);
+    expect(TEST_ROUTE.CHILD.getTypedHash(testHash)).toStrictEqual(undefined);
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedHash(testHash)).toStrictEqual("baz");
 });
 
 it("allows state params parsing", () => {
@@ -910,13 +912,13 @@ it("allows state params parsing", () => {
 
     const state = { foo: "test", bar: 1, untyped: "untyped" };
 
-    expect(TEST_ROUTE.getTypedState(state)).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedState(state)).toEqual({ foo: "test" });
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedState(state)).toEqual({ foo: "test", bar: 1 });
+    expect(TEST_ROUTE.getTypedState(state)).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedState(state)).toStrictEqual({ foo: "test" });
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getTypedState(state)).toStrictEqual({ foo: "test", bar: 1 });
 
-    expect(TEST_ROUTE.getUntypedState(state)).toEqual({ foo: "test", bar: 1, untyped: "untyped" });
-    expect(TEST_ROUTE.CHILD.getUntypedState(state)).toEqual({ bar: 1, untyped: "untyped" });
-    expect(TEST_ROUTE.CHILD.GRANDCHILD.getUntypedState(state)).toEqual({ untyped: "untyped" });
+    expect(TEST_ROUTE.getUntypedState(state)).toStrictEqual({ foo: "test", bar: 1, untyped: "untyped" });
+    expect(TEST_ROUTE.CHILD.getUntypedState(state)).toStrictEqual({ bar: 1, untyped: "untyped" });
+    expect(TEST_ROUTE.CHILD.GRANDCHILD.getUntypedState(state)).toStrictEqual({ untyped: "untyped" });
 });
 
 it("throws if throwable state params are invalid", () => {
@@ -936,8 +938,8 @@ it("throws if throwable state params are invalid", () => {
 
     const state = { foo: "test", bar: "bar", untyped: "untyped" };
 
-    expect(TEST_ROUTE.getTypedState(state)).toEqual({});
-    expect(TEST_ROUTE.CHILD.getTypedState(state)).toEqual({ foo: "test" });
+    expect(TEST_ROUTE.getTypedState(state)).toStrictEqual({});
+    expect(TEST_ROUTE.CHILD.getTypedState(state)).toStrictEqual({ foo: "test" });
     expect(() => TEST_ROUTE.CHILD.GRANDCHILD.getTypedState(state)).toThrow();
 });
 
@@ -985,7 +987,7 @@ it("allows types composition", () => {
 
     assert<IsExact<Parameters<typeof ROUTE.buildState>[0], { fromList?: boolean; hidden?: boolean }>>(true);
 
-    expect(ROUTE.buildPath({ id: 1, subId: 2 }, { page: 1, ordered: true }, "info")).toEqual(
+    expect(ROUTE.buildPath({ id: 1, subId: 2 }, { page: 1, ordered: true }, "info")).toStrictEqual(
         "/1/2?page=1&ordered=true#info"
     );
 
@@ -1002,10 +1004,10 @@ it("allows to trim path pattern", () => {
     assert<IsExact<typeof TEST_ROUTE.CHILD.$.GRANDCHILD.path, "/grand">>(true);
     assert<IsExact<typeof TEST_ROUTE.$.CHILD.GRANDCHILD.path, "/child/grand">>(true);
 
-    expect(TEST_ROUTE.path).toEqual("/test");
-    expect(TEST_ROUTE.$.CHILD.path).toEqual("/child");
-    expect(TEST_ROUTE.CHILD.$.GRANDCHILD.path).toEqual("/grand");
-    expect(TEST_ROUTE.$.CHILD.GRANDCHILD.path).toEqual("/child/grand");
+    expect(TEST_ROUTE.path).toStrictEqual("/test");
+    expect(TEST_ROUTE.$.CHILD.path).toStrictEqual("/child");
+    expect(TEST_ROUTE.CHILD.$.GRANDCHILD.path).toStrictEqual("/grand");
+    expect(TEST_ROUTE.$.CHILD.GRANDCHILD.path).toStrictEqual("/child/grand");
 });
 
 it("allows to inherit non-path params in trimmed children", () => {
@@ -1063,14 +1065,14 @@ it("allows to inherit non-path params in trimmed children", () => {
         >
     >(true);
 
-    expect(TEST_ROUTE.buildPath({}, { foo: 1 }, "hashFoo")).toEqual("/test?foo=1#hashFoo");
-    expect(TEST_ROUTE.$.CHILD.buildPath({}, { foo: 1, bar: "test" }, "hashBar")).toEqual(
+    expect(TEST_ROUTE.buildPath({}, { foo: 1 }, "hashFoo")).toStrictEqual("/test?foo=1#hashFoo");
+    expect(TEST_ROUTE.$.CHILD.buildPath({}, { foo: 1, bar: "test" }, "hashBar")).toStrictEqual(
         "/child?foo=1&bar=test#hashBar"
     );
-    expect(TEST_ROUTE.CHILD.$.GRANDCHILD.buildPath({}, { foo: 1, bar: "test", baz: false }, "hashBaz")).toEqual(
+    expect(TEST_ROUTE.CHILD.$.GRANDCHILD.buildPath({}, { foo: 1, bar: "test", baz: false }, "hashBaz")).toStrictEqual(
         "/grand?foo=1&bar=test&baz=false#hashBaz"
     );
-    expect(TEST_ROUTE.$.CHILD.GRANDCHILD.buildPath({}, { foo: 1, bar: "test", baz: false }, "hashBaz")).toEqual(
+    expect(TEST_ROUTE.$.CHILD.GRANDCHILD.buildPath({}, { foo: 1, bar: "test", baz: false }, "hashBaz")).toStrictEqual(
         "/child/grand?foo=1&bar=test&baz=false#hashBaz"
     );
 
@@ -1102,15 +1104,20 @@ it("prevents path param inheritance in trimmed children", () => {
         true
     );
 
-    expect(TEST_ROUTE.buildPath({ id: 1 })).toEqual("/test/1");
-    expect(TEST_ROUTE.$.CHILD.buildPath({ subId: true })).toEqual("/child/true");
-    expect(TEST_ROUTE.CHILD.$.GRANDCHILD.buildPath({ id: "test" })).toEqual("/grand/test");
-    expect(TEST_ROUTE.$.CHILD.GRANDCHILD.buildPath({ subId: true, id: "test" })).toEqual("/child/true/grand/test");
+    expect(TEST_ROUTE.buildPath({ id: 1 })).toStrictEqual("/test/1");
+    expect(TEST_ROUTE.$.CHILD.buildPath({ subId: true })).toStrictEqual("/child/true");
+    expect(TEST_ROUTE.CHILD.$.GRANDCHILD.buildPath({ id: "test" })).toStrictEqual("/grand/test");
+    expect(TEST_ROUTE.$.CHILD.GRANDCHILD.buildPath({ subId: true, id: "test" })).toStrictEqual(
+        "/child/true/grand/test"
+    );
 
-    expect(TEST_ROUTE.getTypedParams({ id: "1", subId: "true" })).toEqual({ id: 1 });
-    expect(TEST_ROUTE.$.CHILD.getTypedParams({ id: "1", subId: "true" })).toEqual({ subId: true });
-    expect(TEST_ROUTE.CHILD.$.GRANDCHILD.getTypedParams({ id: "1", subId: "true" })).toEqual({ id: "1" });
-    expect(TEST_ROUTE.$.CHILD.GRANDCHILD.getTypedParams({ id: "1", subId: "true" })).toEqual({ id: "1", subId: true });
+    expect(TEST_ROUTE.getTypedParams({ id: "1", subId: "true" })).toStrictEqual({ id: 1 });
+    expect(TEST_ROUTE.$.CHILD.getTypedParams({ id: "1", subId: "true" })).toStrictEqual({ subId: true });
+    expect(TEST_ROUTE.CHILD.$.GRANDCHILD.getTypedParams({ id: "1", subId: "true" })).toStrictEqual({ id: "1" });
+    expect(TEST_ROUTE.$.CHILD.GRANDCHILD.getTypedParams({ id: "1", subId: "true" })).toStrictEqual({
+        id: "1",
+        subId: true,
+    });
 });
 
 it("properly parses arrays in search params", () => {
@@ -1126,7 +1133,7 @@ it("properly parses arrays in search params", () => {
         IsExact<
             ReturnType<typeof TEST_ROUTE.getTypedSearchParams>,
             {
-                a: (number | undefined)[];
+                a: number[];
                 b: number[];
                 e: number[];
             }
@@ -1155,8 +1162,8 @@ it("properly parses arrays in search params", () => {
         b: testArray,
     });
 
-    expect(TEST_ROUTE.getTypedSearchParams(testValue)).toEqual({
-        a: [1, undefined],
+    expect(TEST_ROUTE.getTypedSearchParams(testValue)).toStrictEqual({
+        a: [1],
         b: [],
         e: [1, -1],
     });
@@ -1213,18 +1220,20 @@ it("ensures that default/defined modifiers are applied if a custom type allows u
         >
     >(true);
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ a: "1", b: "2", c: "3" }))).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ a: "1", b: "2", c: "3" }))).toStrictEqual({
         a: 1,
         b: 2,
         c: 3,
     });
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ b: "2", c: "3" }))).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ b: "2", c: "3" }))).toStrictEqual({
+        a: undefined,
         b: 2,
         c: 3,
     });
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ b: "2" }))).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ b: "2" }))).toStrictEqual({
+        a: undefined,
         b: 2,
         c: -1,
     });
@@ -1250,7 +1259,7 @@ it("ensures that arrays don't accept undefined items if a custom type allows und
         IsExact<
             ReturnType<typeof TEST_ROUTE.getTypedSearchParams>,
             {
-                a: (number | undefined)[];
+                a: number[];
                 c: number[];
             }
         >
@@ -1307,7 +1316,7 @@ it("allows to use zod", () => {
         g: { d: testDate },
     });
 
-    expect(plainSearchParams).toEqual({
+    expect(plainSearchParams).toStrictEqual({
         a: "test",
         b: "0",
         c: "false",
@@ -1317,7 +1326,7 @@ it("allows to use zod", () => {
         g: JSON.stringify({ d: testDate }),
     });
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams(plainSearchParams))).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams(plainSearchParams))).toStrictEqual({
         a: "test",
         b: 0,
         c: false,
@@ -1368,7 +1377,7 @@ it("allows to use yup", () => {
         g: { d: testDate },
     });
 
-    expect(plainSearchParams).toEqual({
+    expect(plainSearchParams).toStrictEqual({
         a: "test",
         b: "0",
         c: "false",
@@ -1378,7 +1387,7 @@ it("allows to use yup", () => {
         g: JSON.stringify({ d: testDate }),
     });
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams(plainSearchParams))).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams(plainSearchParams))).toStrictEqual({
         a: "test",
         b: 0,
         c: false,
@@ -1405,11 +1414,11 @@ it("allows to specify defaults within zod types", () => {
         >
     >(true);
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams())).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams())).toStrictEqual({
         a: -1,
     });
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ a: "f" }))).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ a: "f" }))).toStrictEqual({
         a: undefined,
     });
 });
@@ -1430,11 +1439,11 @@ it("allows to specify defaults within yup schemas", () => {
         >
     >(true);
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams())).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams())).toStrictEqual({
         a: -1,
     });
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ a: "f" }))).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ a: "f" }))).toStrictEqual({
         a: undefined,
     });
 });
@@ -1465,13 +1474,13 @@ it("allows to use unions", () => {
         c: true,
     });
 
-    expect(plainSearchParams).toEqual({
+    expect(plainSearchParams).toStrictEqual({
         a: "test",
         b: "1",
         c: "true",
     });
 
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams(plainSearchParams))).toEqual({
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams(plainSearchParams))).toStrictEqual({
         a: "test",
         b: 1,
         c: true,
@@ -1548,53 +1557,20 @@ it("allows to define different types for different route parts", () => {
         >
     >(true);
 
-    expect(TEST_ROUTE.getPlainParams({ id: 1 })).toEqual({ id: "path plain" });
-    expect(TEST_ROUTE.getTypedParams({ id: "" })).toEqual({ id: "path typed" });
-    expect(TEST_ROUTE.getPlainSearchParams({ id: "" })).toEqual({ id: "search plain" });
-    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ id: "" }))).toEqual({ id: -1 });
-    expect(TEST_ROUTE.buildState({ id: new Date() })).toEqual({ id: "state" });
-    expect(TEST_ROUTE.getTypedState({ id: false })).toEqual({ id: false });
+    expect(TEST_ROUTE.getPlainParams({ id: 1 })).toStrictEqual({ id: "path plain" });
+    expect(TEST_ROUTE.getTypedParams({ id: "" })).toStrictEqual({ id: "path typed" });
+    expect(TEST_ROUTE.getPlainSearchParams({ id: "" })).toStrictEqual({ id: "search plain" });
+    expect(TEST_ROUTE.getTypedSearchParams(createSearchParams({ id: "" }))).toStrictEqual({ id: -1 });
+    expect(TEST_ROUTE.buildState({ id: new Date() })).toStrictEqual({ id: "state" });
+    expect(TEST_ROUTE.getTypedState({ id: false })).toStrictEqual({ id: false });
 });
 
 it("generates correct paths when the first segment is optional", () => {
     const TEST_ROUTE = route(":optional?/:required");
 
-    expect(TEST_ROUTE.buildPath({ required: "req" })).toEqual("/req");
-    expect(TEST_ROUTE.buildPath({ optional: "opt", required: "req" })).toEqual("/opt/req");
+    expect(TEST_ROUTE.buildPath({ required: "req" })).toStrictEqual("/req");
+    expect(TEST_ROUTE.buildPath({ optional: "opt", required: "req" })).toStrictEqual("/opt/req");
 
-    expect(TEST_ROUTE.buildRelativePath({ required: "req" })).toEqual("req");
-    expect(TEST_ROUTE.buildRelativePath({ optional: "opt", required: "req" })).toEqual("opt/req");
-});
-
-it("disallows pathname input params when there are no pathname params", () => {
-    const TEST_ROUTE = route("test");
-    const TEST_ROUTE_WITH_EMPTY_PARAMS = route("test", { params: {} });
-
-    // @ts-expect-error There are no pathname params
-    TEST_ROUTE.buildPath({ id: 1 });
-
-    // @ts-expect-error There are no pathname params
-    TEST_ROUTE_WITH_EMPTY_PARAMS.buildPath({ id: 1 });
-});
-
-it("disallows search input params when there are no search params", () => {
-    const TEST_ROUTE = route("test");
-    const TEST_ROUTE_WITH_EMPTY_PARAMS = route("test", { searchParams: {} });
-
-    // @ts-expect-error There are no search params
-    TEST_ROUTE.buildPath({}, { id: 1 });
-
-    // @ts-expect-error There are no search params
-    TEST_ROUTE_WITH_EMPTY_PARAMS.buildPath({}, { id: 1 });
-});
-
-it("disallows state input params when there are no state params", () => {
-    const TEST_ROUTE = route("test");
-    const TEST_ROUTE_WITH_EMPTY_PARAMS = route("test", { state: {} });
-
-    // @ts-expect-error There are no state params
-    TEST_ROUTE.buildState({ id: 1 });
-
-    // @ts-expect-error There are no state params
-    TEST_ROUTE_WITH_EMPTY_PARAMS.buildState({ id: 1 });
+    expect(TEST_ROUTE.buildRelativePath({ required: "req" })).toStrictEqual("req");
+    expect(TEST_ROUTE.buildRelativePath({ optional: "opt", required: "req" })).toStrictEqual("opt/req");
 });